+++ plugin.php
@@ -6,7 +6,7 @@
 $id = 'xlas';
 
 // code version; must be changed for all code changes
-$version = '10.2';
+$version = '10.3';
 
 // sub version must be changes for all changes on resource files
 $sub_version = '1';