+++ src/Views/Data/WriterView.php
@@ -2,52 +2,64 @@
 
 namespace ILIAS\Plugin\LongEssayAssessment\View\Data;
 
+use Edutiek\AssessmentService\Assessment\Data\Location;
+use Edutiek\AssessmentService\Assessment\Data\Writer;
+use Edutiek\AssessmentService\System\Data\UserData;
+use Edutiek\AssessmentService\System\Data\UserDisplay;
+use Edutiek\AssessmentService\Views\Data\ClientSummary;
+use Edutiek\AssessmentService\Views\Data\EssayTaskSummary;
 use Edutiek\AssessmentService\Views\Data\WriterView as WriterViewAbstract;
 
 class WriterView extends WriterViewAbstract
 {
     public function __construct(
-        private readonly \Edutiek\AssessmentService\Assessment\Data\Writer $writer,
-        private readonly \Edutiek\AssessmentService\System\Data\UserData $writer_data,
-        private readonly \Edutiek\AssessmentService\System\Data\UserDisplay $writer_display,
-        private readonly ?\Edutiek\AssessmentService\Assessment\Data\Location $location,
-        private readonly \Edutiek\AssessmentService\Views\Data\EssayTaskSummary $essay_task_summary,
-        private readonly ?\Edutiek\AssessmentService\System\Data\UserData $authorized_by_data,
-        private readonly ?\Edutiek\AssessmentService\System\Data\UserData $excluded_by_data
+        private readonly Writer $writer,
+        private readonly UserData $writer_data,
+        private readonly UserDisplay $writer_display,
+        private readonly ?Location $location,
+        private readonly ClientSummary $client_summary,
+        private readonly EssayTaskSummary $essay_task_summary,
+        private readonly ?UserData $authorized_by_data,
+        private readonly ?UserData $excluded_by_data
     ) {
     }
 
-    public function getWriter(): \Edutiek\AssessmentService\Assessment\Data\Writer
+    public function getWriter(): Writer
     {
         return $this->writer;
     }
 
-    public function getWriterData(): \Edutiek\AssessmentService\System\Data\UserData
+    public function getWriterData(): UserData
     {
         return $this->writer_data;
     }
 
-    public function getWriterDisplay(): \Edutiek\AssessmentService\System\Data\UserDisplay
+    public function getWriterDisplay(): UserDisplay
     {
         return $this->writer_display;
     }
 
-    public function getLocation(): ?\Edutiek\AssessmentService\Assessment\Data\Location
+    public function getLocation(): ?Location
     {
         return $this->location;
     }
 
-    public function getEssayTaskSummary(): \Edutiek\AssessmentService\Views\Data\EssayTaskSummary
+    public function getClientSummary(): ClientSummary
+    {
+        return $this->client_summary;
+    }
+
+    public function getEssayTaskSummary(): EssayTaskSummary
     {
         return $this->essay_task_summary;
     }
 
-    public function getAuthorizedByData(): ?\Edutiek\AssessmentService\System\Data\UserData
+    public function getAuthorizedByData(): ?UserData
     {
         return $this->authorized_by_data;
     }
 
-    public function getExcludedByData(): ?\Edutiek\AssessmentService\System\Data\UserData
+    public function getExcludedByData(): ?UserData
     {
         return $this->excluded_by_data;
     }